+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 2.x branch
 ## 2.4 branch
+### 2.4.5
+* logged users can view future posts, future pages and drafts;
+* the status of a record is properly shown in the admin panel, including publication in the future;
+* added filter for tags;
+* added patterns table for database backups;
+* now you can restore database backups;
+* now you can download and deletes logs.
+
 ### 2.4.4
 * fixed a major bug for the menu builder;
 * fixed bug on log advanced viewer.